+++ common/models/LoginForm.php
@@ -1,7 +1,6 @@
 <?php
-namespace frontend\modules\user\models;
+namespace common\models;
 
-use frontend\models\User;
 use Yii;
 use yii\base\Model;
 
@@ -10,7 +9,7 @@ use yii\base\Model;
  */
 class LoginForm extends Model
 {
-    public $email;
+    public $username;
     public $password;
     public $rememberMe = true;
 
@@ -23,8 +22,8 @@ class LoginForm extends Model
     public function rules()
     {
         return [
-            // email and password are both required
-            [['email', 'password'], 'required'],
+            // username and password are both required
+            [['username', 'password'], 'required'],
             // rememberMe must be a boolean value
             ['rememberMe', 'boolean'],
             // password is validated by validatePassword()
@@ -64,14 +63,14 @@ class LoginForm extends Model
     }
 
     /**
-     * Finds user by [[email]]
+     * Finds user by [[username]]
      *
      * @return User|null
      */
     protected function getUser()
     {
         if ($this->_user === null) {
-            $this->_user = User::findByEmail($this->email);
+            $this->_user = User::findByUsername($this->username);
         }
 
         return $this->_user;
