+++ frontend/config/main.php
@@ -46,10 +46,5 @@ return [
         ],
 
     ],
-    'modules' => [
-        'user' => [
-            'class' => 'frontend\modules\user\Module',
-        ],
-    ],
     'params' => $params,
 ];
