+++ frontend/models/ContactForm.php
@@ -5,7 +5,6 @@ namespace frontend\models;
 use Yii;
 use yii\base\Model;
 
-
 /**
  * ContactForm is the model behind the contact form.
  */
