+++ frontend/models/ResetPasswordForm.php
@@ -1,9 +1,9 @@
 <?php
-namespace frontend\modules\user\models;
+namespace frontend\models;
 
 use yii\base\Model;
 use yii\base\InvalidParamException;
-use frontend\models\User;
+use common\models\User;
 
 /**
  * Password reset form
