+++ frontend/models/SignupForm.php
@@ -1,8 +1,8 @@
 <?php
-namespace frontend\modules\user\models;
+namespace frontend\models;
 
 use yii\base\Model;
-use frontend\models\User;
+use common\models\User;
 
 /**
  * Signup form
