--- frontend/modules/user/Module.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace frontend\modules\user;
-
-/**
- * user module definition class
- */
-class Module extends \yii\base\Module
-{
-    /**
-     * @inheritdoc
-     */
-    public $controllerNamespace = 'frontend\modules\user\controllers';
-
-    /**
-     * @inheritdoc
-     */
-    public function init()
-    {
-        parent::init();
-
-        // custom initialization code goes here
-    }
-}
